Outsource password functions
